Issue #3265: Add helper for creating Middleware that reacts to tabs getting removed.
